requirements.txt
python-allib
environs
langchain
langchain-community
langchain_openai
openai
transformers
xlsxwriter
parsec
pylatex
trinary
torch
torchvision
torchaudio
fuzzysearch